+++ src/plugins/profile/templates/profile.js
@@ -1,6 +1,6 @@
 import 'shared/avatar/avatar.js';
 import { __ } from 'i18n';
-import { _converse, api } from "@converse/headless/core";
+import { api } from "@converse/headless/core";
 import { getPrettyStatus, logOut } from '../utils.js';
 import { html } from "lit";
 
@@ -21,7 +21,6 @@ function tplUserSettingsButton (o) {
 
 export default (el) => {
     const chat_status = el.model.get('status') || 'offline';
-    const fullname = el.model.vcard?.get('fullname') || el.model.vcard?.get('nickname') || _converse.bare_jid;
     const status_message = el.model.get('status_message') || __("I am %1$s", getPrettyStatus(chat_status));
     const i18n_change_status = __('Click to change your chat status');
     const show_settings_button = api.settings.get('show_client_info') || api.settings.get('allow_adhoc_commands');
@@ -44,7 +43,7 @@ export default (el) => {
                         nonce=${el.model.vcard?.get('vcard_updated')}
                         height="40" width="40"></converse-avatar>
                 </a>
-                <span class="username w-100 align-self-center">${fullname}</span>
+                <span class="username w-100 align-self-center">${el.model.getDisplayName()}</span>
                 ${show_settings_button  ? tplUserSettingsButton(el) : ''}
                 ${api.settings.get('allow_logout') ? tplSignout() : ''}
             </div>