+++ CHANGES.md
@@ -7,7 +7,9 @@
 - Fix `autojoin` checkbox state in MUC bookmark form
 - Remove call to `api.confirm` in `@converse/headless`
 - Generate TypeScript declaration files into `dist/types`
+- Removed documentation about the no longer implemented `fullname` option.
 - #3156: Add function to prevent drag stutter effect over iframes when resize is called in overlay mode
+- #3165: Use configured nickname in profile view in the control box
 
 - New config option [stanza_timeout](https://conversejs.org/docs/html/configuration.html#stanza-timeout)